requirements.txt
# Web Framework
fastapi==0.115.12
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
bcrypt==4.3.0

# Validation
pydantic==2.10.6
pydantic-settings==2.7.1
email-validator==2.2.0

# Environment & Configuration
python-dotenv==1.0.1

# Utilities
python-dateutil==2.9.0

# Rate Limiting
slowapi==0.1.9

# WebSocket
websockets==14.1
